// ms-yourdashboard-auth/src/auth/interfaces/auth.interfaces.ts
export interface CustomJwtPayload {
  sub: string; 
  nombre: string;
  sesionId?: string;
  iat?: number;
  exp?: number;
}

// Alias para mayor claridad - ESTA es la que usas en el código
export type JwtPayload = CustomJwtPayload;
export interface UsuarioAutenticado {
  id: string; 
  email: string;
  nombre: string;
  sesion_id: string; 
}